cloudrun/genai/magix/infrastructure/sql/requirements.txt
cloud-sql-python-connector[pg8000]
SQLAlchemy
psycopg2-binary
python-dotenv